requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
braceexpand==0.1.7
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
import_expression==2.2.1.post1
jishaku==2.6.0
joblib==1.4.2
multidict==6.4.3
nltk==3.9.1
propcache==0.3.1
regex==2024.11.6
tabulate==0.9.0
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.13.2
yarl==1.20.0
